Add complex message consumption and update topic name
Updated KafkaConsumerService.cs to include a new method `ConsumeComplexMessageWithIntKey` for consuming messages with an integer key and deserializing them into `OrderCreatedEvent` objects. Adjusted `Program.cs` to change the topic name to "use-case-3-topic" and call the new method. Added a comment in KafkaProducerService.cs about `CustomValueSerializer` and included a `UserId` property in `OrderCreatedEvent`. Introduced `CustomValueDesirializer<T>` for deserialization and added the `OrderCreatedEvent` record.